Enum sp_rpc::list::ListOrValue

pub enum ListOrValue<T> {
    List(Vec<T>),
    Value(T),
}

RPC list or value wrapper.

For some RPCs itโs convenient to call them with either a single value or a whole list of values to get a proper response. In theory you could do a batch query, but itโs:

Less convenient in client libraries

If the response value is small, the protocol overhead might be dominant.

Also itโs nice to be able to maintain backward compatibility for methods that were initially taking a value and now we want to expand them to take a list.

Variants

List(Vec<T>)

A list of values of given type.

Value(T)

A single value of given type.

Implementations

impl<T> ListOrValue<T>

pub fn map<F: Fn(T) -> X, X>(self, f: F) -> ListOrValue<X>

Map every contained value using function F.

This allows to easily convert all values in any of the variants.